refactor: ScoreSequence 임포트 29기로 수정
