+++ frontend/components/interview/modal/score/InterviewScore.tsx
@@ -1,4 +1,4 @@
-import { ScoreSequence } from "@/src/constants/applicant/28";
+import { ScoreSequence } from "@/src/constants/applicant/29";
 import AvgScore from "./AvgScore";
 import InterviewerScores from "./InterviewerScore";
 import { useState } from "react";
